requirements.txt
baidu-aip==2.2.17.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
idna==2.8
imageio==2.6.1
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.4
numpy==1.16.4
opencv-python==3.4.5.20
Pillow==6.2.1
pyparsing==2.4.5
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.22.0
Rtree==0.8.3
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.13.0
tesserocr==2.4.0
urllib3==1.25.3
